Module fedimint::consensus::conflictfilter

Structs

ConflictFilter

The conflict filter is used to ensure that no conflicting transactions are processed in the main loop. If the processing happened sequentially this wouldn’t be a problem, but currently it is done in parallel due to computation intensive operations. This means any conflict could lead to inconsistent outcomes depending on task scheduling.

Traits

ConflictFilterable