Trait fedimint::modules::wallet::bitcoind::BitcoindRpc

pub trait BitcoindRpc: Send + Sync {
    fn get_network<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Network> + Send + 'async_trait, Global>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn get_block_height<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = u64> + Send + 'async_trait, Global>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn get_block_hash<'life0, 'async_trait>(
        &'life0 self,
        height: u64
    ) -> Pin<Box<dyn Future<Output = BlockHash> + Send + 'async_trait, Global>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn get_block<'life0, 'life1, 'async_trait>(
        &'life0 self,
        hash: &'life1 BlockHash
    ) -> Pin<Box<dyn Future<Output = Block> + Send + 'async_trait, Global>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;
    fn get_fee_rate<'life0, 'async_trait>(
        &'life0 self,
        confirmation_target: u16
    ) -> Pin<Box<dyn Future<Output = Option<Feerate>> + Send + 'async_trait, Global>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn submit_transaction<'life0, 'async_trait>(
        &'life0 self,
        transaction: Transaction
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait, Global>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
}

Trait that allows interacting with the Bitcoin blockchain

Functions may panic if if the bitcoind node is not reachable.

Required Methods

fn get_network<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = Network> + Send + 'async_trait, Global>> where
'life0: 'async_trait,
Self: 'async_trait,

Returns the Bitcoin network the node is connected to

fn get_block_height<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = u64> + Send + 'async_trait, Global>> where
'life0: 'async_trait,
Self: 'async_trait,

Returns the current block height

fn get_block_hash<'life0, 'async_trait>(
&'life0 self,
height: u64
) -> Pin<Box<dyn Future<Output = BlockHash> + Send + 'async_trait, Global>> where
'life0: 'async_trait,
Self: 'async_trait,

Returns the block hash at a given height

Panics

If the node does not know a block for that height. Make sure to only query blocks of a height less or equal to the one returned by Self::get_block_height.

While there is a corner case that the blockchain shrinks between these two calls (through on average heavier blocks on a fork) this is prevented by only querying hashes for blocks tailing the chain tip by a certain number of blocks.

fn get_block<'life0, 'life1, 'async_trait>(
&'life0 self,
hash: &'life1 BlockHash
) -> Pin<Box<dyn Future<Output = Block> + Send + 'async_trait, Global>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Returns the block with the given hash

Panics

If the block doesn’t exist.

fn get_fee_rate<'life0, 'async_trait>(
&'life0 self,
confirmation_target: u16
) -> Pin<Box<dyn Future<Output = Option<Feerate>> + Send + 'async_trait, Global>> where
'life0: 'async_trait,
Self: 'async_trait,

Estimates the fee rate for a given confirmation target. Make sure that all federation members use the same algorithm to avoid widely diverging results. If the node is not ready yet to return a fee rate estimation this function returns None.

fn submit_transaction<'life0, 'async_trait>(
&'life0 self,
transaction: Transaction
) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait, Global>> where
'life0: 'async_trait,
Self: 'async_trait,

Submits a transaction to the Bitcoin network

Panics

If the transaction is deemed invalid by the node it was submitted to