Struct fedimint::net::framed::BidiFramed

pub struct BidiFramed<T, WH, RH> {
    sink: FramedSink<WH, T>,
    stream: FramedStream<RH, T>,
}

Framed transport codec for streams

Wraps a stream S and allows sending packetized data of type T over it. Data items are encoded using bincode and the bytes are sent over the stream prepended with a length field. BidiFramed implements Sink<T> and Stream<Item=Result<T, _>>.

Fields
sink: FramedSink<WH, T>stream: FramedStream<RH, T>
Implementations

impl<T, WH, RH> BidiFramed<T, WH, RH> where
WH: AsyncWrite,
RH: AsyncRead,
T: Serialize + DeserializeOwned,

pub fn new<S>(stream: S) -> BidiFramed<T, WriteHalf<S>, ReadHalf<S>> where
S: AsyncRead + AsyncWrite,

Builds a new BidiFramed codec around a stream stream.

See TcpBidiFramed::new_from_tcp for a more efficient version in case the stream is a tokio TCP stream.

pub fn borrow_parts(
&mut self
) -> (&mut FramedSink<WH, T>, &mut FramedStream<RH, T>)

Splits the codec in its sending and receiving parts

This can be useful in cases where potentially simultaneous read and write operations are required. Otherwise a we would need a mutex to guard access.

impl<T> BidiFramed<T, OwnedWriteHalf, OwnedReadHalf> where
T: Serialize + DeserializeOwned,

pub fn new_from_tcp(stream: TcpStream) -> TcpBidiFramed<T>

Special constructor for tokio TCP connections.

Tokio TcpStream implements an efficient method of splitting the stream into a read and a write half this constructor takes advantage of.