Trait fedimint::net::peers::PeerConnections

pub trait PeerConnections<T>where
    T: Serialize + DeserializeOwned + Unpin + Send,{
    fn send<'life0, 'async_trait>(
        &'life0 mut self,
        target: Target<PeerId>,
        msg: T
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn receive<'life0, 'async_trait>(
        &'life0 mut self
    ) -> Pin<Box<dyn Future<Output = (PeerId, T)> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn ban_peer<'life0, 'async_trait>(
        &'life0 mut self,
        peer: PeerId
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    fn into_dyn(self) -> AnyPeerConnections<T>
    where
        Self: Sized + Send + Unpin + 'static,
    { ... }
}

Connection manager that tries to keep connections open to all peers

Production implementations of this trait have to ensure that:

Connections to peers are authenticated and encrypted

Messages are received exactly once and in the order they were sent

Connections are reopened when closed

Messages are cached in case of short-lived network interruptions and resent on reconnect, this avoids the need to rejoin the consensus, which is more tricky.

In case of longer term interruptions the message cache has to be dropped to avoid DoS attacks. The thus disconnected peer will need to rejoin the consensus at a later time.

Required Methods

fn send<'life0, 'async_trait>(
&'life0 mut self,
target: Target<PeerId>,
msg: T
) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>where
'life0: 'async_trait,
Self: 'async_trait,

Send a message to a target, either all peers or a specific one.

The message is sent immediately and cached if the peer is reachable and only cached otherwise.

fn receive<'life0, 'async_trait>(
&'life0 mut self
) -> Pin<Box<dyn Future<Output = (PeerId, T)> + Send + 'async_trait>>where
'life0: 'async_trait,
Self: 'async_trait,

Await receipt of a message from any connected peer.

fn ban_peer<'life0, 'async_trait>(
&'life0 mut self,
peer: PeerId
) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>where
'life0: 'async_trait,
Self: 'async_trait,

Removes a peer connection in case of misbehavior

Provided Methods

fn into_dyn(self) -> AnyPeerConnections<T>where
Self: Sized + Send + Unpin + 'static,

Converts the struct to a PeerConnection trait object